Extend PCI: 64 bit BARs, extended config space, bridges, multiple hosts
- 64 bit BARs: a memory BAR may span two slots and is sized and
  programmed as a register pair; SetBar() takes a 64 bit address. NVMe
  and XHCI declare theirs 64 bit, as their specifications do.
- Host bridges take an optional "mmio64_size" aperture above 4 GB,
  allocated from a second window so several of them may have one.
- 4096 byte configuration space: bus addresses carry a 12 bit register
  number in the ECAM layout, both host bridges route all of it, and
  extended capabilities can be added.
- Devices on a bus marked PCI Express are given a PCI Express
  capability and a device serial number.
- PCI to PCI bridges: new "pci-bridge" device and pci_bridge_init().
  Configuration cycles route by the bus numbers the guest programs,
  INTx is swizzled once per tier, and MSI receivers resolve upward.
- A link carries one device, so a bridge places an upstream port and a
  downstream port per child; the DesignWare host does the same with the
  devices declared under it.
- Multiple host bridges: each gets its own PCI domain, windows and
  interrupt lines. PHYS_MEM_RANGE_MAX raised to 64.
- Fix the status register being treated as write one to clear in full,
  which let a 32 bit write of the command dword clear the capability
  list bit.
- Fix a 32 bit configuration read past the end of configuration space.

+++ vga.cpp
@@ -148,7 +148,7 @@ public:
     uint16_t vbe_regs[VBE_DISPI_INDEX_NB];
 
     void Refresh(SimpleFBDraw *draw) override;
-    void SetBar(int bar_num, uint32_t addr, bool enabled) override;
+    void SetBar(int bar_num, uint64_t addr, bool enabled) override;
 
     uint32_t VbeRead(uint32_t offset, int size_log2);
     void VbeWrite(uint32_t offset, uint32_t val, int size_log2);
@@ -738,7 +738,7 @@ uint32_t VGAState::VbeRead(uint32_t offset, int size_log2)
 }
 
 
-void VGAState::SetBar(int bar_num, uint32_t addr, bool enabled)
+void VGAState::SetBar(int bar_num, uint64_t addr, bool enabled)
 {
     VGAState *s = this;
     if (bar_num == 0)
